Add DHT magnet uri test, udp tracker magnet uri test
